ticket error validation complete

+++ app/api/auth_routes.py
@@ -105,9 +105,12 @@ def delete_user(id):
     Deletes a user account
     """
     user = User.query.filter(User.id == id)
-    user.delete()
-    db.session.commit()
-    return dict(user)
+    if user.username == 'Demo':
+        return {'error': ['You cannot delete the demo user.']}
+    else:
+        user.delete()
+        db.session.commit()
+        return dict(user)
 
 
 @auth_routes.route('/unauthorized')
